+++ changelog.md
@@ -1,5 +1,6 @@
 # Version 2024.11.3 (2024-11-18)
 
+- Add interface(\_id) to performance log message
 - Reduce MAX_CACHE_AGE to 15s
 
 # Version 2024.11.2 (2024-11-17)